Catch OSError when JSON string is passed to Path
I've tried to be as surgical as possible, as I want the bug fix to be
backwards compatible.

It's likely that an OSError is raised when a JSON string is passed to
Path, rather than when the name component of the path is too long for
the file system; i.e. the design of the function, and not real-world
use, led to the bug. Consequently, we catch an OSError before
opening/parsing.

We could parse before opening, in two nested try/except blocks. However,
if parsing failed, then we would not be able to report as much
information to the user.